Executing checkov on GitHub Actions workflows only (#293)
When activating checkov through GitHub Advanced Security for kustomize
files, I noticed so many alerts created (which is good), but 99% of
those did not seemed to be worth reporting (IMHO).

Failing to find the ability to configure the severity or the tool
itself, reconfiguring it with trunk to analyse GitHub Actions workflows
only for now. (Those were good).