PTFE-1144 gcloud instances support of fixed labels (#496)
This pull request includes changes related to the configuration and
testing of custom labels for GCP instances in the runner manager
backend. The most important changes include adding a new attribute to
the `GCPInstanceConfig` class to configure custom labels, and modifying
the GCP backend test to verify the correct setting of custom labels.