PTFE-1798 handling errors of no such index (#617)
If the redis instance happens to reboot during the runtime of the runner
manager, redis will need for the `Migrator().run()` function to run to
re-create the indexes. Otherwise we will face an interruption of service
with the error:

```shell
redis.exceptions.ResponseError: runner-manager:runner_manager.models.runner_group.RunnerGroup:index: no such index
```

The following actions were taken in the code:
- Reduce the index interval from one hour to 15 minutes.
- Add more context to the indexing function, with this information.
- When failing to contact the redis instance, on the healtcheck,
schedule a index run for a more dynamic self-healing approach.