aws-backend: add support for running instances on multiple subnets (#700)
Update the AWS backend (and instance config) to add the ability to run
instance on a set of subnets (and
so availability zone).

The main use-case for this is to allow instance to come up in multiple
availability zones. A user would create one (or more) subnets in an
availability zone and the list the subnets in the `subnet_configs`
section. This has two main advantages:
- There is some crude load balancing (subnets are picked at random from
the list)
- If an instance fails to start we can retry in another subnet. This
will help in cases where an AZ has been starved of all the instance and
instance not longer start.

To support this (without breaking existing configs) I added a new
`subnet_configs` section to the instance config. This is mutually
exclusive to the existing `subnet_id` section.
If `subnet_configs` is set, then the backend will choose one of the
subnets at random and try to bring up the instance there. If this fails.
then another subnet will be tried, until one works. If none of the
subnets work, an the exception will be re-raised.